Solved some more linter issues and moved somneo class into seperate file
